% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.pogsnet}
\alias{coef.pogsnet}
\title{pogsnet coefficients.}
\usage{
\method{coef}{pogsnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{Value returned by pogsnet(..).}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Returns coeficients of pogsnet fit.
}

