% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.pogsnet}
\alias{plot.pogsnet}
\title{pogsnet coefficient plot.}
\usage{
\method{plot}{pogsnet}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{pogsnet object.}

\item{xvar}{x-variable to plot coefficient against.}

\item{...}{Extra parameters for matplot.}
}
\description{
Plots coefficients
}

