% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.pogsnet}
\alias{predict.pogsnet}
\title{Pogsnet prediction}
\usage{
\method{predict}{pogsnet}(object, newx, s = NULL, type = c("link",
  "response", "class"), ...)
}
\arguments{
\item{object}{Value returned by pogsnet(..).}

\item{newx}{New value of x for which prediction should be done.}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{type}{Predict response or class.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Predict response for new value input
}

